Fix Reminder time inputs losing typed digits; extract formatTimestamp
The hour/minute fields' :value bound straight to timeHour/timeMinute
with raw el.value mutation for sanitizing. Vue's DOM patcher compares
the actual el.value against its own tracked value on every re-render of
that vnode (not just when the bound value changes) and force-corrects
on mismatch — so a keystroke in one field got silently reverted the
instant anything else in the component re-rendered (e.g. the other
field committing on blur when focus moved between them). This broke
the "schedules its reminder path" and "near-future reminder becomes
Focus" e2e tests: minute stayed stuck at its untouched default instead
of the typed value.

Routes the fields through their own draft refs (hourDraft/minuteDraft)
that Vue's own reactivity fully owns, so it never needs to "correct"
the DOM out from under a driver/user's typed input. Commit-on-blur
behavior (and the two-digit-entry fix from 763130f) is unchanged;
onDone also flushes the drafts directly now instead of via template
refs. Verified locally: full reminder-flow.spec.ts +
reminder-notification-actions.spec.ts (8/8) now pass.

Also extracts formatTimestamp (Today/Yesterday/date + time formatting)
from QuestListItem.vue into src/utils/timestamp.ts as a shared utility,
and drops the redundant local git-permission override (already covered
globally).
