+++ src/utils/timestamp.ts
@@ -1,0 +1,11 @@
+/** Formats an ISO timestamp as "Today, HH:MM", "Yesterday, HH:MM", or "Mon D, HH:MM". */
+export function formatTimestamp(raw: string): string {
+  const date = new Date(raw);
+  const today = new Date();
+  const yesterday = new Date(today);
+  yesterday.setDate(today.getDate() - 1);
+  const time = date.toLocaleTimeString("en-US", { hour: "2-digit", minute: "2-digit", hour12: false });
+  if (date.toDateString() === today.toDateString()) return `Today, ${time}`;
+  if (date.toDateString() === yesterday.toDateString()) return `Yesterday, ${time}`;
+  return date.toLocaleDateString("en-US", { month: "short", day: "numeric" }) + `, ${time}`;
+}
